requirements.txt
openai
python-dotenv
snowflake-snowpark-python[pandas]
snowflake-snowpark-python
pytest